requirements.txt
# Telegram
python-telegram-bot==21.*

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# LangChain + LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# OpenAI
openai>=1.30.0

# Google APIs
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0

# Vector store
faiss-cpu>=1.8.0

# Web search
tavily-python>=0.3.0

# LangSmith tracing
langsmith>=0.1.0

# Supabase (eval logging / agent_logs table)
supabase>=2.4.0

# Utilities
python-dotenv>=1.0.0
